update akka 2.5.32, use built-in shard passivation instead of rolling our own (#139)
This reverts most of #136, with the exception of the configuration plumbing.

akka 2.5 already has a built-in mechanism for automated actor passivation. It will also handle buffering messages to the actor in case it is resurrected (e.g. if the timeout fires but it is given another message).

This required updating akka to the latest patch release (2.5.16 -> 2.5.32). As a consequence a few deprecated calls were updated, notably PatternsCS -> Patterns.